no float

+++ ACpower.cpp
@@ -135,8 +135,8 @@ void ACpower::control()
 	{
 		uint16_t Pold;
 		_zero++;
-		Inow = sqrt((float)_I2summ / _Icntr) * _Iratio;
-		Unow = sqrt((float)_U2summ / _Ucntr) * _Uratio;  // if Uratio !=1 требуется изменение схемы и перекалибровка подстроечником!
+		Inow = sqrt(_I2summ / _Icntr) * _Iratio;
+		Unow = sqrt(_U2summ / _Ucntr) * _Uratio;  // if Uratio !=1 требуется изменение схемы и перекалибровка подстроечником!
 		Pold = Pavg;
 		Pavg = Pnow;
 		Pnow = Inow * Unow;
